requirements.txt
click==7.1.2
config==0.5.0.post0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
peewee==3.13.3
PyYAML==5.3.1
Werkzeug==1.0.1
